FN-5668
As an Fenix1 user, I need to be able to create a Poste even if its REF
is larger than an integer

+++ focDataSource_DB/src/com/foc/focDataSourceDB/db/DBManagerServer.java
@@ -299,7 +299,7 @@ public class DBManagerServer {
 			    	  	}
 			    	  	
 			    	  	while(rs.next()){
-			    	  		int ref = rs.getInt(indexOfREF);
+			    	  		long ref = rs.getLong(indexOfREF);
 	
 			    	  		focObject.setReference_WithFocListRefHashMapAdjustment(ref);
 			    	  	}
